ai: <pasted ci error>
response:

It looks like you're encountering an error when running the black formatter on your project. The error message says that the currently activated Python version (3.8.14) is not supported by the project and that it is trying to find and use a compatible version (3.10.6), but it is unable to import the typing_extensions module.

There are a few possible solutions to this problem:

You can try using a different version of Python that is compatible with the project. You can specify the version of Python you want to use by adding a python-version field to your pyproject.toml file, like this:
Copy code
[tool.poetry]
...
python-version = 3.10
This will tell Poetry to use Python 3.10 when installing and running your project's dependencies.

If you want to continue using Python 3.8.14, you can try installing the typing_extensions module manually. You can do this by running the following command:
Copy code
poetry run pip install typing_extensions
This should install the typing_extensions module and allow the black formatter to run properly.

If the above solutions don't work, you can try running the black formatter with the --skip-string-normalization flag, like this:
Copy code
poetry run black . --check --skip-string-normalization
This will tell the black formatter to skip normalizing string literals, which may allow it to run successfully even if the typing_extensions module is not available.